# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Medicamentos",
        "description": "API para gerenciamento de medicamentos e lembretes",
        "contact": {
            "name": "Equipe de Suporte",
            "email": "suporte@exemplo.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/medication-reminder": {
            "get": {
                "tags": [
                    "Medication Reminders"
                ],
                "summary": "Listar lembretes de medicamentos",
                "description": "Retorna lista de lembretes, podendo filtrar por medication_id",
                "operationId": "ec60e7d3097796e52a682da92dd7a0e6",
                "parameters": [
                    {
                        "name": "paginate",
                        "in": "query",
                        "description": "Retornar paginado (true/false)",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": "true"
                        }
                    },
                    {
                        "name": "medication_id",
                        "in": "query",
                        "description": "ID do medicamento para filtrar",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de lembretes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MedicationReminder"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Medication Reminders"
                ],
                "summary": "Criar novo lembrete",
                "description": "Cria um novo lembrete de medicamento",
                "operationId": "f0785b62c0bf2038fa47722b60f3472a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MedicationReminderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lembrete criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MedicationReminder"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/medication-reminder/{id}": {
            "get": {
                "tags": [
                    "Medication Reminders"
                ],
                "summary": "Obter lembrete específico",
                "description": "Retorna detalhes de um lembrete pelo ID",
                "operationId": "26311962757a76e48dea816f8fc9a466",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do lembrete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do lembrete",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MedicationReminder"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lembrete não encontrado"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Medication Reminders"
                ],
                "summary": "Atualizar lembrete",
                "description": "Atualiza um lembrete existente",
                "operationId": "d85f87b8d89d156ca749e6e75d4596d2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do lembrete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MedicationReminderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Lembrete atualizado com sucesso"
                    },
                    "422": {
                        "description": "Erro ao atualizar",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Lembrete não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Medication Reminders"
                ],
                "summary": "Excluir lembrete",
                "description": "Remove um lembrete existente",
                "operationId": "25d26abd829d56c01cacabe5d5308874",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do lembrete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Lembrete excluído com sucesso"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Lembrete não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Medication": {
                "title": "Medication",
                "description": "Medication model representation",
                "required": [
                    "id",
                    "name",
                    "user_id",
                    "presentation"
                ],
                "properties": {
                    "id": {
                        "description": "ID do medicamento",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "Nome do medicamento",
                        "type": "string",
                        "example": "Paracetamol"
                    },
                    "user_id": {
                        "description": "ID do usuário dono do medicamento",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "presentation": {
                        "description": "Forma de apresentação do medicamento",
                        "type": "string",
                        "enum": [
                            "SOLVENT",
                            "CAPSULE",
                            "TABLET"
                        ],
                        "example": "CAPSULE"
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Data de atualização",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00Z"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "reminders": {
                        "description": "Lembretes associados ao medicamento",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MedicationReminder"
                        }
                    }
                },
                "type": "object"
            },
            "MedicationReminder": {
                "title": "Medication Reminder",
                "description": "Modelo de lembrete para medicamentos",
                "required": [
                    "dose",
                    "user_id",
                    "hour",
                    "medication_id",
                    "reminder_type"
                ],
                "properties": {
                    "id": {
                        "description": "ID do lembrete",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "dose": {
                        "description": "Dose a ser tomada",
                        "type": "string",
                        "example": "1 comprimido"
                    },
                    "user_id": {
                        "description": "ID do usuário",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "week_day": {
                        "description": "Dia da semana (para lembretes semanais)",
                        "type": "string",
                        "enum": [
                            "SUNDAY",
                            "MONDAY",
                            "TUESDAY",
                            "WEDNESDAY",
                            "THURSDAY",
                            "FRIDAY",
                            "SATURDAY"
                        ],
                        "example": "MONDAY",
                        "nullable": true
                    },
                    "quantity": {
                        "description": "Quantidade da dose",
                        "type": "number",
                        "format": "float",
                        "example": 1.5,
                        "nullable": true
                    },
                    "hour": {
                        "description": "Hora do lembrete",
                        "type": "string",
                        "format": "time",
                        "example": "08:00"
                    },
                    "reminder_type": {
                        "$ref": "#/components/schemas/ReminderType"
                    },
                    "date": {
                        "description": "Data específica (para lembretes de data específica)",
                        "type": "string",
                        "format": "date",
                        "example": "2023-12-31",
                        "nullable": true
                    },
                    "medication_id": {
                        "description": "ID do medicamento relacionado",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Data de atualização",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00Z"
                    },
                    "medication": {
                        "$ref": "#/components/schemas/Medication"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "User model representation",
                "required": [
                    "id",
                    "name",
                    "password",
                    "birth_date"
                ],
                "properties": {
                    "id": {
                        "description": "ID do medicamento",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "Nome",
                        "type": "string",
                        "example": "Paracetamol"
                    },
                    "email": {
                        "description": "Email",
                        "type": "string",
                        "example": "example@email.com"
                    },
                    "password": {
                        "description": "Senha do usuário",
                        "type": "string",
                        "example": "example@email.com"
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Data de atualização",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "ReminderType": {
                "description": "Tipo de lembrete: AGENDADO (recorrente) ou ÚNICO (data específica)",
                "type": "string",
                "enum": [
                    "SCHEDULED",
                    "SINGLE"
                ],
                "example": "SCHEDULED"
            },
            "MedicationReminderRequest": {
                "title": "Medication Reminder Request",
                "required": [
                    "dose",
                    "hour",
                    "medication_id",
                    "reminder_type"
                ],
                "properties": {
                    "dose": {
                        "type": "string",
                        "example": "1 comprimido"
                    },
                    "week_day": {
                        "type": "string",
                        "enum": [
                            "SUNDAY",
                            "MONDAY",
                            "TUESDAY",
                            "WEDNESDAY",
                            "THURSDAY",
                            "FRIDAY",
                            "SATURDAY"
                        ],
                        "example": "MONDAY",
                        "nullable": true
                    },
                    "quantity": {
                        "type": "number",
                        "format": "float",
                        "example": 1.5,
                        "nullable": true
                    },
                    "hour": {
                        "type": "string",
                        "format": "time",
                        "example": "08:00"
                    },
                    "reminder_type": {
                        "$ref": "#/components/schemas/ReminderType"
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-12-31",
                        "nullable": true
                    },
                    "medication_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "prefix": "Bearer "
            }
        }
    },
...
}